Большой апгрейд
1) Перестроил структуру труб и КС, теперь это настоящие классы, с методами и т.д.
2) Теперь консоль это класс только со статическими методами, а класса проверок больше нет, но зато есть пространство имён, которое содержит все проверки.
3) Сделал адекватное удаление, теперь всё работает так как надо.
4) Отредактировал код в некоторых местах, сделал его более логичным и читабельным.

+++ WorkOfBadamshin/WorkOfBadamshin/Console.cpp
@@ -47,7 +47,7 @@ void Console::PrintErrorText(std::string textError)
 }
 
 // Изменение шрифта консоли
-void Console::ChangeConsoleFont(size_t width, size_t height)
+void Console::ChangeConsoleFont(short width, short height)
 {
 	CONSOLE_FONT_INFOEX cfi;
 	cfi.cbSize = sizeof(cfi);
